separate construction_mpi
