+++ src/projects/hpcspades/pipeline.cpp
@@ -28,8 +28,7 @@
 
 #include "stages/genomic_info_filler.hpp"
 #include "stages/read_conversion.hpp"
-#include "stages/construction.hpp"
-#include "stages/construction_mpi.hpp"
+#include "common/stages/construction_mpi.hpp"
 #include "stages/simplification.hpp"
 #include "stages/ss_edge_split.hpp"
 #include "configs/config_struct.hpp"
@@ -188,14 +187,11 @@ static void AddSimplificationStages(StageManager &SPAdes) {
         SPAdes.add<debruijn_graph::SSEdgeSplit>();
 }
 
-static void AddConstructionStages(StageManager &SPAdes, bool mpi = false) {
+static void AddConstructionStages(StageManager &SPAdes) {
     using namespace debruijn_graph::config;
     pipeline_type mode = cfg::get().mode;
 
-    if (mpi)
-        SPAdes.add<debruijn_graph::ConstructionMPI>();
-    else
-        SPAdes.add<debruijn_graph::Construction>();
+    SPAdes.add<debruijn_graph::ConstructionMPI>();
     if (!PipelineHelper::IsMetagenomicPipeline(mode))
         SPAdes.add<debruijn_graph::GenomicInfoFiller>();
 }
@@ -270,7 +266,7 @@ void assemble_genome(bool mpi = false) {
     SPAdes->add<ReadConversion>();
 
     if (!AssemblyGraphPresent()) {
-        AddConstructionStages(*SPAdes, mpi);
+        AddConstructionStages(*SPAdes);
         if (cfg::get().sewage)
             SPAdes->add<debruijn_graph::RestrictedEdgesFilling>();
